lolp base trait
