+++ src/solver/variable/CMakeLists.txt
@@ -121,7 +121,6 @@ set(SRC_VARIABLE_ECONOMY
         include/antares/solver/variable/economy/loldCsr.h
         include/antares/solver/variable/economy/lolp.h
         include/antares/solver/variable/economy/lolp_base.h
-        include/antares/solver/variable/economy/lolpCsr.h
         include/antares/solver/variable/economy/nearPriceCap.h
         include/antares/solver/variable/economy/avail-dispatchable-generation.h
         include/antares/solver/variable/economy/dispatchable-generation-margin.h
