+++ src/solver/variable/include/antares/solver/variable/economy/lolp_base.h
@@ -1,4 +1,3 @@
-
 /*
  * Copyright 2007-2025, RTE (https://www.rte-france.com)
  * See AUTHORS.txt
@@ -21,3 +20,34 @@
  */
 
 #pragma once
+
+#include "economy_base.h"
+
+namespace Antares::Solver::Variable::Economy
+{
+
+struct LOLP_Base_Traits
+{
+    static std::string Unit()
+    {
+        return "%";
+    }
+
+    typedef Results<R::AllYears::Average<>> ResultsType;
+
+    static constexpr uint8_t decimal = 2;
+
+    static constexpr uint8_t spatialAggregate = Category::spatialAggregateOr;
+
+    static double value()
+    {
+        return 100.;
+    }
+
+    static void computeStats(IntermediateValues& iv)
+    {
+        iv.computeStatisticsOrForTheCurrentYear();
+    }
+};
+
+} // namespace Antares::Solver::Variable::Economy
